refactor(evolution): fold promotion plans into audit
